Fixed null pointer error
Error occurred when trying to run a method on the stage object
applicationStage. I didn't realize I needed to create a Stage since it
is not done in the main loop (but running the java program itself
presents a stage as an argument!)
I referenced
https://docs.oracle.com/javase/8/javafx/api/javafx/stage/Stage.html to
figure out how to fix it!