Copied study menu code/files over to data/deck versions.
I copied the code from the openstudymenu event to have versions for the
data and deck menus. Added fxml files and controllers also copied from
the study menu. I then changed the key words in each file to match the
respective menu. The skeletons for the menus are now implemented. These
are all the implementations I originally imagined for the main menu.